// src/lib/api/generated/models/linkDTO.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Thisiscz API
 * Thisiscz 后端 API 文档
 * OpenAPI spec version: v1
 */
import type { LinkCategory } from './linkCategory'
import type { UserDTO } from './userDTO'

export interface LinkDTO {
  id?: number
  createdAt?: string
  updatedAt?: string
  title?: string
  description?: string
  url?: string
  imageUrl?: string
  category?: LinkCategory
  userId?: string
  user?: UserDTO
}